refined table generator for exporting local batch info
