+++ src/utils/utils.py
@@ -1,4 +1,6 @@
 import yaml
+import logging
+from tqdm import tqdm
 
 def read_yaml(path: str) -> dict:
     """Reads a YAML file and returns its content as a dictionary."""
@@ -8,3 +10,12 @@ def read_yaml(path: str) -> dict:
         return data
     except Exception as e:
         raise FileNotFoundError(f"File does not exist : {path}")
+
+class TqdmLoggingHandler(logging.StreamHandler):
+    def emit(self, record):
+        try:
+            msg = self.format(record)
+            tqdm.write(msg)
+            self.flush()
+        except Exception:
+            self.handleError(record)
